Alteração da classe Gerente, valores monetários de double para BigDecimal

+++ src/main/java/modulobonus/Gerente.java
@@ -1,8 +1,10 @@
 package modulobonus;
 
+import java.math.BigDecimal;
+
 public class Gerente extends Funcionario{
     
-    public Gerente(int id, double salario, Departamento departamento) {
+    public Gerente(int id, BigDecimal salario, Departamento departamento) {
         super(id, salario, Cargo.GERENTE, departamento);
     }
